Update beatmaps as a stress test.
More difficult beatmaps will have an framerate issue as there are too many objects on the screen at once. In order to reduce the number of objects, we need to:
- Make the notes faster
- Spawn the notes sooner.

Considering the latter, otherwise I may need a new implementation for detecting when notes are close to the hitbar (consider using rTP ratio between 0.9 and 1.1 or something like that).