Add ability to spawn and score drag notes (see below)
At the moment i'm not sure how I want to judge the accurace of a drag note. A possible solution would be to judge how long they were 'on path' for relative to the note, normalize that number to Length * NoteHelper.Quarter, and divide the players score with the total possible score.